Add receiving events tests to VK channel
